+++ packages/react/src/components/RichText.tsx
@@ -3,6 +3,8 @@ import remarkGfm from 'remark-gfm';
 import { MemoizedReactMarkdown } from './MemoizedReactMarkdown';
 import rehypeRaw from 'rehype-raw';
 import { useConfig } from '@opencx/widget-react-headless';
+import { Dialoger, DialogerContent } from './Dialoger';
+import { ZoomableImage } from './ZoomableImage';
 
 export function RichText({
   children,
@@ -27,6 +29,28 @@ export function RichText({
             <a target={props.target || anchorTarget || '_top'} {...props}>
               {children}
             </a>
+          );
+        },
+        img: ({ src, alt, ...props }) => {
+          if (!src) return <img src={src} alt={alt} {...props} />;
+          return (
+            <Dialoger
+              trigger={
+                <img
+                  src={src}
+                  alt={alt}
+                  {...props}
+                  className="cursor-pointer rounded-xl"
+                />
+              }
+            >
+              <DialogerContent
+                className="size-full max-w-full rounded-3xl flex items-center justify-center bg-transparent border-none gap-0"
+                withClose
+              >
+                <ZoomableImage src={src} alt={alt} />
+              </DialogerContent>
+            </Dialoger>
           );
         },
       }}